RoommateSync_Backend_v3/backend/RoommateSync_v2/roommate_sync_upgraded/requirements.txt
# ── Core Framework ─────────────────────────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0
flask-mail>=0.10.0

# ── WebSockets (NEW) ───────────────────────────────────────────────────────────
flask-socketio>=5.3.0
python-socketio>=5.10.0

# ── Rate Limiting (NEW) ────────────────────────────────────────────────────────
flask-limiter>=3.5.0
limits>=3.6.0

# ── Database ───────────────────────────────────────────────────────────────────
pymongo>=4.6.0

# ── Auth ───────────────────────────────────────────────────────────────────────
PyJWT>=2.8.0
bcrypt>=4.1.0

# ── ML / Data Science (NEW matching engine) ────────────────────────────────────
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0

# ── Computer Vision / OCR ──────────────────────────────────────────────────────
opencv-python-headless>=4.9.0
pytesseract>=0.3.10
Pillow>=10.2.0

# ── Environment ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
gunicorn>=22.0.0
eventlet>=0.36.0
